requirements.txt
Pillow
python-barcode
pywin32
ttkbootstrap
